Add application mode
