+++ src/apps/ocioview/ocioview/viewer/__init__.py
@@ -1,6 +1,8 @@
 # SPDX-License-Identifier: BSD-3-Clause
 # Copyright Contributors to the OpenColorIO Project.
 
-from .image_viewer import ViewerChannels, ImageViewer
+from .base_image_viewer import ViewerChannels, BaseImageViewer
+from .edit_image_viewer import EditImageViewer
 from .offscreen_viewer import WgpuCanvasOffScreenViewer
+from .preview_image_viewer import PreviewImageViewer
 from .utils import load_image
